LITERAS/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
aiohttp==3.9.1
websockets==12.0
python-multipart==0.0.6
autogen-ext==0.4.0.dev11
autogen-agentchat==0.4.0.dev11
